LayoutTests/fast/dom/HTMLSelectElement/select-add-optgroup-expected.txt

Test select.add() with optgroup

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS deepCopy() is "0,1,2,Y1"
PASS deepCopy() is "0,1,2,Y2,Y1"
PASS deepCopy() is "0,1,2,Y3,Y2,Y1"
PASS deepCopy() is "0,1,2,Y4,Y3,Y2,Y1"
PASS deepCopy() is "0,1,2,X,Y"
PASS deepCopy() is "0,1,X,Y,2"
PASS deepCopy() is "0,1,X,Y,X,Y,2"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS deepCopy() is "0,1,2,Y"
PASS mySelect.add("foo") is undefined.
PASS mySelect.add("foo", 0) is undefined.
PASS mySelect.add(undefined) is undefined.
PASS mySelect.add(undefined, 0) is undefined.
PASS mySelect.add(null) is undefined.
PASS mySelect.add(null, 0) is undefined.
PASS successfullyParsed is true

TEST COMPLETE
